aws_requirements.txt
requests==2.9.1
PyJWT==1.7.1